Usermanagement_Service/Usermanagement/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
Django==5.1.2
django-appconf==1.0.6
django-celery==3.1.17
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cors-headers==4.5.0
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.37.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
pillow==10.3.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
cookiecutter==2.6.0
django-structlog==9.1.1
python-json-logger==3.3.0
prometheus-client==0.21.1
structlog==25.2.0